+++ src/Ext/TechnoType/Body.h
@@ -24,12 +24,12 @@ public:
 		PhobosFixedString<0x20> GroupAs;
 		Valueable<int> RadarJamRadius;
 		Valueable<int> InhibitorRange;
-		Valueable<double> MindControlRangeLimit;
+		Valueable<Leptons> MindControlRangeLimit;
 		Valueable<bool> Interceptor;
-		Valueable<double> Interceptor_GuardRange;
-		Valueable<double> Interceptor_MinimumGuardRange;
-		Valueable<double> Interceptor_EliteGuardRange;
-		Valueable<double> Interceptor_EliteMinimumGuardRange;
+		Valueable<Leptons> Interceptor_GuardRange;
+		Valueable<Leptons> Interceptor_MinimumGuardRange;
+		Valueable<Leptons> Interceptor_EliteGuardRange;
+		Valueable<Leptons> Interceptor_EliteMinimumGuardRange;
 		Valueable<CoordStruct> TurretOffset;
 		Valueable<bool> Powered_KillSpawns;
 		Valueable<bool> Spawn_LimitedRange;
@@ -97,12 +97,12 @@ public:
 			GroupAs(NONE_STR),
 			RadarJamRadius(0),
 			InhibitorRange(0),
-			MindControlRangeLimit(-1.0),
+			MindControlRangeLimit(),
 			Interceptor(false),
-			Interceptor_GuardRange(0.0),
-			Interceptor_MinimumGuardRange(0.0),
-			Interceptor_EliteGuardRange(0.0),
-			Interceptor_EliteMinimumGuardRange(0.0),
+			Interceptor_GuardRange(),
+			Interceptor_MinimumGuardRange(),
+			Interceptor_EliteGuardRange(),
+			Interceptor_EliteMinimumGuardRange(),
 			TurretOffset({0, 0, 0}),
 			Powered_KillSpawns(false),
 			Spawn_LimitedRange(false),